+++ PMC Reader/UIColor+PHColor.m
@@ -20,47 +20,47 @@
 @implementation UIColor (PHColor)
 
 + (UIColor *)backgroundColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHBackgroundColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)cellBackgroundColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHCellBackgroundColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)iconColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHIconColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)textColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHTextColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)linkColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHLinkColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)popoverBackgroundColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHPopoverBackgroundColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)popoverButtonColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHPopoverButtonColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)popoverBorderColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     return [kPHPopoverBorderColorArray objectAtIndex:backgroundIndex];
 }
 
 + (UIColor *)popoverIconColor {
-    int backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
+    NSInteger backgroundIndex =[[NSUserDefaults standardUserDefaults] integerForKey:@"Background"];
     if (backgroundIndex == 2) {
         return kPHNightTextColor;
     }
